🐛 google/fb id from int to bigint

+++ database/migrations/20190606182938_Users.js
@@ -11,8 +11,8 @@ exports.up = function (knex, Promise) {
         tbl.string("email", 128).notNullable();
         tbl.string("phone_num", 128);
         tbl.string("password", 16);
-        tbl.integer("google_id").unsigned();
-        tbl.integer("facebook_id").unsigned();
+        tbl.bigInteger("google_id").unsigned();
+        tbl.bigInteger("facebook_id").unsigned();
       })
   );
 };
